refactor(vended-tools): move stop tool to experimental (#3465)
