+++ site/src/content/docs/user-guide/concepts/tools/vended-tools-imports.ts
@@ -48,5 +48,5 @@ import { makeSleep } from '@strands-agents/sdk/vended-tools/sleep'
 
 // --8<-- [start:stop_import]
 import { Agent } from '@strands-agents/sdk'
-import { stop } from '@strands-agents/sdk/vended-tools/stop'
+import { stop } from '@strands-agents/sdk/experimental/vended-tools/stop'
 // --8<-- [end:stop_import]
