+++ site/src/content/docs/user-guide/concepts/tools/vended-tools.mdx
@@ -11,10 +11,10 @@ sourceLinks:
   - path: strands-ts/src/vended-tools/http-request/http-request.ts
   - path: strands-ts/src/vended-tools/notebook/notebook.ts
   - path: strands-ts/src/vended-tools/sleep/sleep.ts
-  - path: strands-ts/src/vended-tools/stop/stop.ts
+  - path: strands-ts/src/experimental/vended-tools/stop/stop.ts
   - path: strands-py/src/strands/vended_tools/http_request/http_request.py
   - path: strands-py/src/strands/vended_tools/sleep/sleep.py
-  - path: strands-py/src/strands/vended_tools/stop/stop.py
+  - path: strands-py/src/strands/experimental/tools/stop/stop.py
 ---
 
 Vended tools are pre-built tools included directly in the Strands SDK for common agent tasks like file operations, shell commands, HTTP requests, and persistent notes. 
@@ -40,7 +40,7 @@ Each tool is imported from its own subpath under `@strands-agents/sdk/vended-too
 | [Notebook](#notebook) | Manage persistent text notebooks | TypeScript (Node.js, browsers) |
 | [Bash](#bash) | Execute shell commands with persistent sessions | Python, TypeScript (Node.js, Unix/Linux/macOS) |
 | [Sleep](#sleep) | Pause execution for a bounded, cancellable duration | Python, TypeScript (Node.js, browsers) |
-| [Stop](#stop) | Gracefully end the agent loop when the task is complete | Python, TypeScript (Node.js, browsers) |
+| [Stop](#stop-experimental) | Gracefully end the agent loop when the task is complete | Python, TypeScript (Node.js, browsers) |
 
 ### File Editor
 
@@ -269,7 +269,10 @@ agent = Agent(tools=[short_sleep])
 
 ---
 
-### Stop
+### Stop (Experimental)
+
+> This tool is experimental and subject to change in future
+> revisions without notice.
 
 Lets the model gracefully end the agent loop with an optional final message. The default loop already terminates when the model returns without any tool call; the stop tool is useful when you want an explicit "I am done" affordance, when a workflow enforces that termination is a deliberate model decision, or when a sub-agent needs to signal completion back to a coordinator via the loop's last assistant message.
 
@@ -295,7 +298,7 @@ The two SDKs shim onto different loop-termination primitives, which produces a s
 
 ```python
 from strands import Agent
-from strands.vended_tools import stop
+from strands.experimental.tools import stop
 
 agent = Agent(
     tools=[stop],
@@ -307,7 +310,7 @@ result = agent("Summarize the changes in ./CHANGELOG.md")
 </Tab>
 </Tabs>
 
-📖 [Full API Reference](https://github.com/strands-agents/harness-sdk/blob/main/strands-ts/src/vended-tools/stop/README.md)
+📖 [Full API Reference](https://github.com/strands-agents/harness-sdk/blob/main/strands-ts/src/experimental/vended-tools/stop/README.md)
 
 ---
