+++ strands-py/src/strands/experimental/tools/stop/__init__.py
@@ -1,9 +1,11 @@
 """Tool for gracefully ending the agent loop.
 
+This tool is experimental and subject to change in future revisions without notice.
+
 Example Usage:
     ```python
     from strands import Agent
-    from strands.vended_tools import stop
+    from strands.experimental.tools import stop
 
     agent = Agent(tools=[stop])
     ```
