+++ strands-py/src/strands/vended_tools/__init__.py
@@ -1,4 +1,4 @@
-"""Built-in tools for commands, files, HTTP, pausing, and agent-loop control.
+"""Built-in tools for commands, files, HTTP, and pausing.
 
 The :data:`bash` tool runs a
 persistent shell on the host; the :func:`make_bash` and :func:`make_file_editor`
@@ -6,8 +6,6 @@ factories produce sandbox-routed tools that either bind to a
 :class:`~strands.sandbox.base.Sandbox` at creation (as the built-in Docker/SSH
 sandboxes do when vending tools) or read the sandbox from the agent at call time.
 The :data:`sleep` tool pauses execution for a bounded, cancellable duration.
-The :data:`stop` tool signals the event loop to end the current invocation
-gracefully.
 
 The :data:`http_request` tool makes raw HTTP calls with a strict default
 security posture (private-network denial, redirect and body-size caps,
@@ -17,9 +15,9 @@ controls when needed.
 Example Usage:
     ```python
     from strands import Agent
-    from strands.vended_tools import bash, file_editor, http_request, sleep, stop
+    from strands.vended_tools import bash, file_editor, http_request, sleep
 
-    agent = Agent(tools=[bash, file_editor, http_request, sleep, stop])
+    agent = Agent(tools=[bash, file_editor, http_request, sleep])
     ```
 """
 
@@ -27,7 +25,6 @@ from .bash import bash, make_bash
 from .file_editor import file_editor, make_file_editor
 from .http_request import http_request, make_http_request
 from .sleep import make_sleep, sleep
-from .stop import make_stop, stop
 
 __all__ = [
     "bash",
@@ -37,7 +34,5 @@ __all__ = [
     "make_file_editor",
     "make_http_request",
     "make_sleep",
-    "make_stop",
     "sleep",
-    "stop",
 ]
