+++ strands-ts/src/experimental/vended-tools/stop/index.ts
