+++ strands-ts/src/experimental/vended-tools/stop/types.ts
