+++ strands-ts/src/vended-tools/index.ts
@@ -3,7 +3,7 @@
  *
  * Provides a single import path for consumers who want all built-in tools:
  * ```typescript
- * import { bash, fileEditor, httpRequest, notebook, sleep, stop } from '@strands-agents/sdk/vended-tools'
+ * import { bash, fileEditor, httpRequest, notebook, sleep } from '@strands-agents/sdk/vended-tools'
  * ```
  *
  * Note: This module requires a Node.js environment because the `bash` tool
@@ -16,4 +16,3 @@ export * from './file-editor/index.js'
 export * from './http-request/index.js'
 export * from './notebook/index.js'
 export * from './sleep/index.js'
-export * from './stop/index.js'
